Update build system to use webpack for 3DViewer
The webpage for the 3D viewer is now connected to the same webpack configuration as the rest of the application.

Reworked a fair amount of the js for the 3D viewer to make it more modern and fix some scoping issues.
