+++ .github/workflows/deploy.yml
@@ -23,7 +23,7 @@ jobs:
       - run: sudo apt-get install -y libappindicator1 fonts-liberation ruby-full python3-pip
       - run: sudo pip install mkdocs
       - run: npm ci
-      - run: npm run openstudio-build && sed -i 's/\/static/static/g' dist/index.html
+      - run: npm run openstudio-build && sed -i 's/\/static/static/g' dist/index.html && sed -i 's/\/static/..\/static/g' dist/3DViewer/index.html
       - run: mkdocs build -d dist/docs
       - run: cp -r icons/ dist/ && cp site.webmanifest dist/ && cp browserconfig.xml dist/
       - name: Deploy 🚀
