--- 3DViewer/index.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<title>Preview</title>
-<meta charset="UTF-8" />
-<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
-<script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r71/three.min.js"></script>
-<script src="https://cdn.rawgit.com/mrdoob/three.js/r71/examples/js/controls/OrbitControls.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/dat-gui/0.5/dat.gui.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/1.16.1/TweenLite.min.js"></script>
-</head>
-<body>
-<style type="text/css">
-#heads-up {
-  background-color:#ddd;
-  border-radius:8px;
-  display:none;
-  padding:0 5px 10px;
-  opacity:0.85;
-  position:absolute;
-  left:50px;
-  text-align:left;
-  z-index:10;
-}
-#color-bar {
-  border:1px solid #777;
-  position:absolute;
-  bottom:20px;
-  left:20px;
-  pointer-events:all;
-}
-#color-bar-container {
-  display:none;
-  pointer-events:none;
-  position:absolute;
-  transition:height 0.5s ease, transform 0.5s ease, top 0.5s ease, left 0.5s ease;
-}
-.ignore-transition {
-  transition:height 0.5s ease, transform 0.5s ease !important;
-}
-.color-bar-tick {
-  background-color:#777;
-  height:15px;
-  width:2px;
-  position:absolute;
-}
-.color-bar-text {
-  transform:rotate(300deg);
-  transform-origin:top left;
-  display:inline-block;
-  position:absolute;
-  color:#777;
-  transition: transform 0.5s ease;
-}
-.rotate {
-  height:1px !important; /* Fix for dragging when rotated */
-  transform:rotate(90deg);
-}
-.rotate .color-bar-text {
-  transform:rotate(270deg);
-}
-</style>
-<script type='text/javascript'>
-
-var Module = {
-  preRun: [],
-  postRun: [],
-  canvas: (function() {
-    var canvas = document.getElementById('canvas');
-    return canvas;
-  })(),
-  totalDependencies: 0,
-  monitorRunDependencies: function(left) {
-    this.totalDependencies = Math.max(this.totalDependencies, left);
-  }
-};
-
-</script>
-<script src='./libtest_lib.js'></script>
-<script src='./index.js'></script>
-</body>
-</html>
